requirements.txt
langchain==0.3.4
langchain-aws==0.2.4
langchain-chroma==0.1.4
langchain-community==0.3.3
langchain-ollama==0.2.0
chromadb==0.5.15
pypdf==5.0.1
fastapi==0.115.3
uvicorn==0.32.0
python-multipart==0.0.12
black==24.10.0
python-dotenv==1.0.1